+++ packages/espressocash_app/lib/features/outgoing_link_payments/data/repository.dart
@@ -246,7 +246,7 @@ extension on OutgoingLinkPayment {
         txFailureReason: status.toTxFailureReason(),
         cancelTx: status.toCancelTx(),
         cancelTxId: status.toCancelTxId(),
-        slot: status.toSlot().toString(),
+        slot: status.toSlot()?.toString(),
         generatedLinksAt: linksGeneratedAt,
         resolvedAt: status.toResolvedAt(),
         publicKey: await toPublicKey(),