+++ packages/solana/lib/src/rpc/dto/simple_instruction.g.dart
@@ -13,3 +13,10 @@ SimpleInstruction _$SimpleInstructionFromJson(Map<String, dynamic> json) =>
       accounts:
           (json['accounts'] as List<dynamic>).map(AccountKey.fromJson).toList(),
     );
+
+Map<String, dynamic> _$SimpleInstructionToJson(SimpleInstruction instance) =>
+    <String, dynamic>{
+      'programId': instance.programId,
+      'data': instance.data,
+      'accounts': instance.accounts.map((e) => e.toJson()).toList(),
+    };